se agrega rutas de las carnes como navegacion post productos
