--- src/app/components/cerdo/cerdo.component.html
@@ -1,15 +1,0 @@
-<div>
-  <input type="text" [(ngModel)]="searchQuery" (input)="onSearch()" placeholder="Buscar por nombre de corte...">
-</div>
-
-<section>
-  <div class="articles-container">
-    <article class="card" *ngFor="let corte of filteredCortes">
-      <h2>{{ corte.nombre }}</h2>
-      <p>Preparaciones:</p>
-      <ul>
-        <li *ngFor="let preparacion of corte.preparaciones">{{ preparacion }}</li>
-      </ul>
-    </article>
-  </div>
-</section>
