--- src/app/components/pollo/pollo.component.css
@@ -1,43 +1,0 @@
-/* src/app/components/pollo/pollo.component.css, src/app/components/cerdo/cerdo.component.css, src/app/components/vacuno/vacuno.component.css */
-
-section {
-  margin: 0 auto;
-}
-
-section article {
-  transition: all 0.4s ease;
-  background-color: #fff;
-  margin: 1rem 0;
-  border-radius: 0.5rem;
-  padding: 0.25rem 0;
-}
-
-section article:hover {
-  box-shadow: 0 0.1rem 2rem rgba(0,0,0,0.3);
-}
-
-section article h2, section article p {
-  margin: 1rem;
-}
-
-.articles-container {
-  display: flex;
-  flex-direction: row; /* Alinear horizontalmente */
-  gap: 1rem; /* Espacio entre artículos */
-  justify-content: space-around;
-}
-
-.card {
-  position: relative;
-  cursor: pointer; /* Hace que el cursor cambie a una mano indicando que es clicable */
-  padding: 1rem;
-  border: 1px solid #ccc; /* Añadir un borde para resaltar las tarjetas */
-  border-radius: 8px; /* Bordes redondeados */
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Añadir una sombra sutil */
-  background-color: #fff; /* Fondo blanco */
-  transition: transform 0.3s; /* Añadir una transición para un efecto de hover */
-}
-
-.card:hover {
-  transform: scale(1.02); /* Efecto de hover */
-}
