+++ src/app/components/productos/productos.component.html
@@ -1,124 +1,79 @@
 <html>
-  <header
-    class="d-flex flex-wrap align-items-center justify-content-center justify-content-md-between py-0 mb-0 border-bottom"
-  >
-    <div class="col-md-3 mb-2">
-      <a
-        href="/dashboard"
-        class="d-inline-flex link-body-emphasis text-decoration-none"
-      >
-        <img
-          class="logo-duarte-letras"
-          src="../../../assets/img/logo-duarte-letras.png"
-          alt="logo-duarte-letras"
-      /></a>
-    </div>
-
-    <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-      <li><a href="/inicio" class="nav-link px-2">Inicio</a></li>
-      <li>
-        <a href="/productos" class="nav-link px-2 link-secondary">Productos</a>
-      </li>
-      <li><a href="/mis-pedidos" class="nav-link px-2">Mis Pedidos</a></li>
-      <li><a href="/mis-datos" class="nav-link px-2">Mis Datos</a></li>
-      <li>
-        <a href="/sobre-nosotros" class="nav-link px-2">Sobre Nosotros</a>
-      </li>
-    </ul>
-
-    <div class="col-md-3 text-end">
-      <button type="button" class="btn btn-outline-primary">
-        Cerrar Sesión
-      </button>
-    </div>
-  </header>
-  <main>
-    <div class="container">
-      <!-- Título: Quiénes Somos -->
-      <div class="row">
-        <div class="col text-center">
-          <h1>Carnes Duarte</h1>
-        </div>
-      </div>
-
-      <!-- Título: Tradición Familiar -->
-      <div class="row mt-1">
-        <div class="col text-center">
-          <h2>Tradición Familiar</h2>
-        </div>
-      </div>
-      <section>
-        <div class="articles-container">
-          <article class="card" (click)="navigateTo('vacuno')">
-            <h2>Vacuno</h2>
-            <p>...</p>
-            <img src="../../../assets/img/img-duarte/vacuno2.png" width="400" />
-            <p>Esto es un vacuno</p>
-          </article>
-          <article class="card" (click)="navigateTo('pollo')">
-            <h2>Pollo</h2>
-            <p>...</p>
-            <img src="../../../assets/img/img-duarte/pollo.png" width="400" />
-            <p>Esto es un pollo</p>
-          </article>
-          <article class="card" (click)="navigateTo('cerdo')">
-            <h2>Cerdo</h2>
-            <p>...</p>
-            <img src="../../../assets/img/img-duarte/cerdo.png" width="400" />
-            <p>Esto es un cerdo</p>
-          </article>
-        </div>
-      </section>
-    </div>
-  </main>
-  <footer class="py-5 my-0">
-    <div class="container" style="background-color: rgba(206, 0, 25, 0.6)">
-      <div class="row align-items-center">
-        <div class="col-md-3 mb-2 mb-md-0">
-          <a
-            href="/inicio"
-            class="d-inline-flex link-body-emphasis text-decoration-none"
-          >
-            <img
-              class="logo-duarte-letras"
-              src="../../../assets/img/logo-duarte-letras.png"
-              alt="logo-duarte-letras"
-            />
-          </a>
-        </div>
-        <div class="col-md-9">
-          <ul class="nav justify-content-center border-bottom pb-3 mb-3">
-            <li class="nav-item">
-              <a href="/inicio" class="nav-link px-2 text-body-secondary"
-                >Inicio</a
-              >
-            </li>
-            <li class="nav-item">
-              <a href="/productos" class="nav-link px-2 link-secondary"
-                >Productos</a
-              >
-            </li>
-            <li class="nav-item">
-              <a href="/mis-pedidos" class="nav-link px-2 text-body-secondary"
-                >Mis Pedidos</a
-              >
-            </li>
-            <li class="nav-item">
-              <a href="/mis-datos" class="nav-link px-2 text-body-secondary"
-                >Mis Datos</a
-              >
-            </li>
-            <li class="nav-item">
-              <a
-                href="/sobre-nosotros"
-                class="nav-link px-2 text-body-secondary"
-                >Sobre Nosotros</a
-              >
-            </li>
-          </ul>
-          <p class="text-center text-body-secondary">©2024 Carnes Duarte</p>
-        </div>
+<header class="d-flex flex-wrap align-items-center justify-content-center justify-content-md-between py-0 mb-0 border-bottom">
+  <div class="col-md-3 mb-2">
+    <a href="/dashboard" class="d-inline-flex link-body-emphasis text-decoration-none">
+      <img class="logo-duarte-letras" src="../../../assets/img/logo-duarte-letras.png" alt="logo-duarte-letras"/>
+    </a>
+  </div>
+  <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
+    <li><a href="/inicio" class="nav-link px-2">Inicio</a></li>
+    <li><a href="/productos" class="nav-link px-2 link-secondary">Productos</a></li>
+    <li><a href="/mis-pedidos" class="nav-link px-2">Mis Pedidos</a></li>
+    <li><a href="/mis-datos" class="nav-link px-2">Mis Datos</a></li>
+    <li><a href="/sobre-nosotros" class="nav-link px-2">Sobre Nosotros</a></li>
+  </ul>
+  <div class="col-md-3 text-end">
+    <button type="button" class="btn btn-outline-primary">Cerrar Sesión</button>
+  </div>
+</header>
+<main>
+  <div class="container">
+    <!-- Título: Quiénes Somos -->
+    <div class="row">
+      <div class="col text-center">
+        <h1>Carnes Duarte</h1>
       </div>
     </div>
-  </footer>
+    <!-- Título: Tradición Familiar -->
+    <div class="row mt-1">
+      <div class="col text-center">
+        <h2>Tradición Familiar</h2>
+      </div>
+    </div>
+    <section>
+      <div class="articles-container">
+        <article class="card" (click)="navigateTo('vacuno')">
+          <h2>Vacuno</h2>
+          <p>...</p>
+          <img src="../../../assets/img/img-duarte/vacuno2.png" width="400"/>
+          <p>Esto es un vacuno</p>
+        </article>
+        <article class="card" (click)="navigateTo('pollo')">
+          <h2>Pollo</h2>
+          <p>...</p>
+          <img src="../../../assets/img/img-duarte/pollo.png" width="400"/>
+          <p>Esto es un pollo</p>
+        </article>
+        <article class="card" (click)="navigateTo('cerdo')">
+          <h2>Cerdo</h2>
+          <p>...</p>
+          <img src="../../../assets/img/img-duarte/cerdo.png" width="400"/>
+          <p>Esto es un cerdo</p>
+        </article>
+      </div>
+    </section>
+    <router-outlet></router-outlet>
+  </div>
+</main>
+<footer class="py-5 my-0">
+  <div class="container" style="background-color: rgba(206, 0, 25, 0.6)">
+    <div class="row align-items-center">
+      <div class="col-md-3 mb-2 mb-md-0">
+        <a href="/inicio" class="d-inline-flex link-body-emphasis text-decoration-none">
+          <img class="logo-duarte-letras" src="../../../assets/img/logo-duarte-letras.png" alt="logo-duarte-letras"/>
+        </a>
+      </div>
+      <div class="col-md-9">
+        <ul class="nav justify-content-center border-bottom pb-3 mb-3">
+          <li class="nav-item"><a href="/inicio" class="nav-link px-2 text-body-secondary">Inicio</a></li>
+          <li class="nav-item"><a href="/productos" class="nav-link px-2 link-secondary">Productos</a></li>
+          <li class="nav-item"><a href="/mis-pedidos" class="nav-link px-2 text-body-secondary">Mis Pedidos</a></li>
+          <li class="nav-item"><a href="/mis-datos" class="nav-link px-2 text-body-secondary">Mis Datos</a></li>
+          <li class="nav-item"><a href="/sobre-nosotros" class="nav-link px-2 text-body-secondary">Sobre Nosotros</a></li>
+        </ul>
+        <p class="text-center text-body-secondary">©2024 Carnes Duarte</p>
+      </div>
+    </div>
+  </div>
+</footer>
 </html>
